am64x: kernel: Fix dhrystone benchmark overflow bug
- Modify data type of variable used for storing Dhrystones per second
  from uint32_t to float.
- Add typecasting to calculation of Dhrystones per second so result is
  also float.
- Update results for single-core & SMP Dhrystone benchmark.

Fixes: MCUSDK-12108

+++ examples/kernel/freertos/dhrystone_benchmark/dhrystone_benchmark.c
@@ -47,7 +47,7 @@ void dhrystone_benchmark_main(void *args)
     Drivers_open();
     Board_driversOpen();
 
-    uint32_t dhryPerSec = 0;
+    float dhryPerSec = 0;
     int status = SystemP_SUCCESS;
 
     /* Run Dhrystone demo with 1 thread */
@@ -56,7 +56,7 @@ void dhrystone_benchmark_main(void *args)
     {
         DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Iterations", DHRY_ITERATIONS);
         DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Threads", 1);
-        DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Dhrystones per second", dhryPerSec);
+        DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %6.1f\r\n", "Dhrystones per second", dhryPerSec);
         DebugP_log("\r\n");
     }
 
@@ -68,7 +68,7 @@ void dhrystone_benchmark_main(void *args)
         {
             DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Iterations", DHRY_ITERATIONS);
             DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Threads", 2);
-            DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Dhrystones per second", dhryPerSec);
+            DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %6.1f\r\n", "Dhrystones per second", dhryPerSec);
             DebugP_log("\r\n");
         }
     }
@@ -81,7 +81,7 @@ void dhrystone_benchmark_main(void *args)
         {
             DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Iterations", DHRY_ITERATIONS);
             DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Threads", 5);
-            DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Dhrystones per second", dhryPerSec);
+            DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %6.1f\r\n", "Dhrystones per second", dhryPerSec);
             DebugP_log("\r\n");
         }
     }
@@ -94,7 +94,7 @@ void dhrystone_benchmark_main(void *args)
         {
             DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Iterations", DHRY_ITERATIONS);
             DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Threads", 10);
-            DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %d\r\n", "Dhrystones per second", dhryPerSec);
+            DebugP_log("[DHRYSTONE BENCHMARKING] %-32s : %6.1f\r\n", "Dhrystones per second", dhryPerSec);
             DebugP_log("\r\n");
         }
     }
